requirements.txt
psycopg2==2.9.1
tqdm==4.62.1
APScheduler==3.7.0
requests==2.26.0
python-dotenv==0.19.0
pandas==1.3.2
pyarrow==5.0.0